rpc: Rename parsers in to parser.
